requirements.txt
pydantic-ai
python-dotenv
gradio
qdrant-client
datasets
sentence-transformers